+++ Main.Tests/Instructions Execution/LD (IX+d),n + LD (IY+d),n     .Tests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
-using Ploeh.AutoFixture;
+using AutoFixture;
+
 
 namespace Konamiman.Z80dotNet.Tests.InstructionsExecution
 {
@@ -26,7 +27,7 @@ namespace Konamiman.Z80dotNet.Tests.InstructionsExecution
 
             Execute(opcode, prefix, offset, newValue);
 
-            Assert.AreEqual(newValue, ProcessorAgent.Memory[actualAddress]);
+            Assert.AreEqual(newValue, (int)ProcessorAgent.Memory[actualAddress]);
         }
 
         [Test]
