+++ Main.Tests/Instructions Execution/LD I,A + LD I,R           .Tests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
-using Ploeh.AutoFixture;
+using AutoFixture;
+
 
 namespace Konamiman.Z80dotNet.Tests.InstructionsExecution
 {
@@ -24,7 +25,7 @@ namespace Konamiman.Z80dotNet.Tests.InstructionsExecution
 
             Execute(opcode, prefix);
 
-            Assert.AreEqual(newValue, GetReg<byte>(reg));
+            Assert.AreEqual(newValue, (int)GetReg<byte>(reg));
         }
 
         [Test]
