+++ Main.Tests/Instructions Execution/LD r,(rr)       .Tests.cs
@@ -1,5 +1,6 @@
 ﻿using NUnit.Framework;
-using Ploeh.AutoFixture;
+using AutoFixture;
+
 
 namespace Konamiman.Z80dotNet.Tests.InstructionsExecution
 {
@@ -35,7 +36,7 @@ namespace Konamiman.Z80dotNet.Tests.InstructionsExecution
 
             Sut.Execute(opcode);
 
-            Assert.AreEqual(newValue, GetReg<byte>(destReg));
+            Assert.AreEqual(newValue, (int)GetReg<byte>(destReg));
         }
 
         [Test]
